added new line of code

+++ src/main/java/ie/marlowe/calculator.java
@@ -2,7 +2,7 @@ package ie.marlowe;
 
 public class calculator {
     // This method will add two numbers.
-    public double add(double num1, double num2){
+    public int add(int num1, int num2){
         return num1 + num2;
     }
 }
